refactor(decompose): split webapp/e2e-tests/electron/for_feature_development_not_LT_verification/terminals/content/electron-inject-badge-in-titlebar.spec.ts into folder
